Moved home page to /app/ uri. Consolidated new files under their respective parent directory.
