+++ src/main/java/com/pse/fotoz/spring/MvcConfig.java
@@ -24,9 +24,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter
     "com.pse.fotoz.controllers.producer",
     "com.pse.fotoz.controllers.customers",
     "com.pse.fotoz.controllers.photographers",
-    "com.pse.fotoz.controllers.common",
-    "com.pse.fotoz.controllers.home",
-    "com.pse.fotoz.controllers.logout",})
+    "com.pse.fotoz.controllers.common"})
 @EnableWebMvc
 public class MvcConfig extends WebMvcConfigurerAdapter {
