+++ src/main/webapp/WEB-INF/views/common/logout/logout.twig
@@ -7,7 +7,7 @@
 {% block more_css %}
     <link href="/assets/common/css/default.css" rel="stylesheet"/>
     <link href="/assets/common/css/bootstrap/sb-admin/font-awesome/css/font-awesome.min.css" rel="stylesheet"/>
-    <link href="/assets/logout/css/logout.css" rel="stylesheet"/>
+    <link href="/assets/common/css/logout.css" rel="stylesheet"/>
 {% endblock %}
 
 {% block body %}
