+++ src/main/webapp/WEB-INF/views/customers/home/index.twig
@@ -8,7 +8,7 @@
     <link href="/assets/common/css/default.css" rel="stylesheet"/>
 
 
-    <link href="/assets/home/css/shop-homepage.css" rel="stylesheet">
+    <link href="/assets/customers/css/shop-homepage.css" rel="stylesheet">
 {% endblock %}
 
 {% block body %}
@@ -78,13 +78,13 @@
                             </ol>
                             <div class="carousel-inner">
                                 <div class="item active">
-                                    <img class="slide-image" src="/assets/home/img/nature1.jpg" alt="">
+                                    <img class="slide-image" src="/assets/customers/img/nature1.jpg" alt="">
                                 </div>
                                 <div class="item">
-                                    <img class="slide-image" src="/assets/home/img/nature2.jpg" alt="">
+                                    <img class="slide-image" src="/assets/customers/img/nature2.jpg" alt="">
                                 </div>
                                 <div class="item">
-                                    <img class="slide-image" src="/assets/home/img/nature3.jpg" alt="">
+                                    <img class="slide-image" src="/assets/customers/img/nature3.jpg" alt="">
                                 </div>
                             </div>
                             <a class="left carousel-control" href="#carousel-example-generic" data-slide="prev">
@@ -257,12 +257,6 @@
 
     </div>
     <!-- /.container -->
-
-    <!-- jQuery -->
-    <script src="/assets/home/js/jquery.js"></script>
-
-    <!-- Bootstrap Core JavaScript -->
-    <script src="/assets/home/js/bootstrap.min.js"></script>
 
     {% include "common/snippets/language_bar.twig" %}
