+++ src/main/webapp/assets/common/css/logout.css
